Various improvements and nitpicks

+++ colorer/src/Colorer-library/src/colorer/unicode/CharacterClass.cpp
@@ -254,7 +254,7 @@ void CharacterClass::addCategory(ECharCategory cat)
 void CharacterClass::addCategory(const String &cat)
 {
   for (size_t pos = 0; pos < ARRAY_SIZE(char_category_names); pos++) {
-    int ci;
+    size_t ci;
     for (ci = 0; ci < cat.length() && cat[ci] == char_category_names[pos][ci]; ci++);
     if (ci == cat.length()) addCategory(ECharCategory(pos));
   }
@@ -283,7 +283,7 @@ void CharacterClass::clearCategory(ECharCategory cat)
 void CharacterClass::clearCategory(const String &cat)
 {
   for (size_t pos = 0; pos < ARRAY_SIZE(char_category_names); pos++) {
-    int ci;
+    size_t ci;
     for (ci = 0; ci < cat.length() && cat[ci] == char_category_names[pos][ci]; ci++);
     if (ci == cat.length()) clearCategory(ECharCategory(pos));
   }
